make Remote a final class
