+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - `Innmind\OperatingSystem\Sockets::connectTo()` now returns an `Innmind\Immutable\Attempt`
 - `Innmind\OperatingSystem\OperatingSystem` is now a final class
 - `Innmind\OperatingSystem\Sockets` is now a final class
+- `Innmind\OperatingSystem\Remote` is now a final class
 
 ### Fixed
